feat: 시니어-청년 상세정보
